# Optimize `fast_string_to_time`